requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
requests==2.26.0
python-dotenv==0.19.2
loguru==0.5.3